Add `Bitmap` type for representing texture bitmap data in a renderer-agnostic form
This is meant as a vehicle for passing texture bitmap data into the renderer's image cache (where it will presumably be converted into a renderer-specific form and moved over to the GPU). It is roughly equivalent to `SkBitmap`, but using Ink types only, so that it doesn't need to be used with Skia.

PiperOrigin-RevId: 673915143