Handle travel direction more explicitly in the `BrushTipModeler` and helpers.
This change replaces relying on the direction of the modeled velocity with calculating the average travel direction based on the previous and next input positions, if they exist. The direction is handled as an optional property for cases where the inputs are not moving or there is only a single input so far. This way, the cases do not depend on the behavior of `atan2(0, 0)`. And the direction does not depend on the input modeler's velocity output.

Because travel direction depends on the next input's position, the `BrushTipModeler` tracks when brush behaviors depend on this information and only use up to the second-to-last "stable" modeled input for "fixed" tip states. As a result, if the first modeled input was first handled on its own, it will get modeled again when subsequent inputs are added to the stroke. This takes care of the sharp discontinuity caused by the travel direction being undefined for the first input until the second input exists.

PiperOrigin-RevId: 678808654