Handle invalid Append more gently in InProgressStroke
The three calls to StrokeInputBatch::Append in InProgressStroke (append the real inputs to the queued real inputs, append the queued real inputs to the processed inputs, subsequently append the predicted inputs to the processed inputs) all should be guaranteed to pass after the validation done by ValidateNewInputs at the start of EnqueueInputs.

Crashing might be too harsh if that assumption is violated, though. While that does make some bugs more visible, Append does do its validation up-front, so the consequences of missing validation done on Append are:

* If the new inputs could not be appended to the queued real inputs, no consequence.

* If the new inputs could be appended to the queued real inputs, but the first of those could not be appended to the processed input, then the previous prediction might be incorrectly cleared (will be cleared if this is the end of the inputs or there are some real or predicted inputs queued).

* If the prediction couldn't be appended to the processed inputs, that will be skipped.

Instead of CHECK-failing in those cases, log an error and return an invalid status.

PiperOrigin-RevId: 680684944