Rename `kPi` to `kHalfTurn`, and similarly `kTwoPi` and `kHalfPi`
The primary reason for doing this is for naming consistency with our [Jetpack API](https://android.googlesource.com/platform/frameworks/support/+/c40905d87f290be25cdc25454f5dd46ff47e4613/ink/ink-geometry/src/jvmAndroidMain/kotlin/androidx/ink/geometry/Angle.kt#66).

The other reason is principle: the `Angle` type is designed to abstract away its units--that is, an `Angle` object is an opaque representation of an angle, and you have to call a method if you want to get a representation of that angle as e.g. a number of radians or a number of degrees--so it doesn't really make sense to have a constant of type `Angle` named `kPi`.  It would make more sense to name it something like `kPiRadians` or `k180Degrees` or `kHalfTurn`.  This CL adopts the last of these names (again, for consistency with Jetpack).

An additional benefit: renaming `::ink::kPi` (which is an `Angle`) removes the ambiguity with `::ink::numbers::kPi` (which is a `double`).

PiperOrigin-RevId: 684044741