Add `uTextureMapping` uniform to shader to distinguish between tiling and winding textures
The "local" texture coordinates returned by our fragment shader need to be calculated differently depending on whether we are using a tiling or winding texture.  This CL adds a new uniform (and necessary surrounding plumbing) so that the shader can know which texture mapping mode to use.  A subsequent CL will make use of this uniform to actually perform the necessary calculations.

PiperOrigin-RevId: 693070926