Add `textureCoords` varying and use `uTextureMapping` uniform to decide how to calculate it
Previously the return value of the fragment shader was always `varyings.position`, but now we instead return the new `varyings.textureCoords`, which is calculated in the vertex shader.  When `uTextureMapping` is 0 (i.e. `BrushPaint::TextureMapping::kTiling`), then `varyings.textureCoords` is set to `varyings.position`, so we get the same behavior as before.  However, when `uTextureMapping` is 1 (i.e. `BrushPaint::TextureMapping::kWinding`), then `varyings.textureCoords` is instead set to `attributes.surfaceUv`.

In the case that the surface UV attribute is packed, the new `unpackSurfaceUv` function will decode it.  This function is based on that attribute using `MeshFormat::AttributeType::kFloat2PackedIn4UnsignedBytes_X12_Y20`.  Recall that Skia will encode each of the four unsigned bytes (0-255) as a half float (0-1); this function is basically doing bit manipulation in the form of floating point multiplications, similarly to the other shader functions in that file.

PiperOrigin-RevId: 693851140