Fix `NoiseGenerator` to use all 64 seed bits, not just the bottom 32.
This was an oversight in the original implementation of `NoiseGenerator`.  It turns out that while `std::seed_seq` will happily accept 64-bit entries, it just ignores all but the bottom 32 bits.  ಠ_ಠ  If you want to use 64 seed bits, you have to provide two separate 32-bit entries.

Unfortunately, this change does break backwards seed compatibility for `NoiseGenerator`, which we normally don't want to ever do.  If we needed another fix like this farther down the road, we'd need to do more work to make backwards compatibility possible, but fortunately nobody is relying on this yet, so it's not too late to just fix this the easy way.

PiperOrigin-RevId: 718498101