Add per-stroke random seed values for use with `BrushBehavior::NoiseNode`
This CL allows supplying `InProgressStroke` with a `stroke_seed` value that is used to help seed the random number generators for `NoiseNode`s in the brush, if any.  These means that two strokes with the same randomized brush will (in general) use different random values.

To prevent existing strokes from changing when serialized or regenerated, we must also persist that each dry stroke's seed somehow.  The most expedient places to keep it are within the `Brush` or in the `StrokeInputBatch` (since our API is built to assume that these two things combined are sufficient to regenerate a stroke).  This CL opts to store this value in the `StrokeInputBatch`, since that's the component that already varies from stroke to stroke.

PiperOrigin-RevId: 721042476