Add `BrushBehavior::Target::kAnimationFrameOffset`
This CL adds the necessary boilerplate and plumbing for a new brush behavior target for controlling the starting animation frame of animated particles.  Such a target can be wired up to e.g. a noise node to make particles start on a random frame, or a distance-traveled source to make a series of particles that cycle through starting frames in order, or whatever else is desired.

The actual extruder change to use the new field in the `BrushTipState` to alter the particle's texture UV coordinates is left as a TODO for the next CL.

PiperOrigin-RevId: 721741202