Add mesh attribute for per-particle animation offset
The high-level idea here is that, for a particle-based mesh, we pack surface UV into 3 bytes (12 bits for each of U and V), and use the 4th byte of the GPU attribute to store 8 bits of per-particle animation offset.  For an extruded mesh, we pack surface UV into 4 bytes (12 bits for U and 20 bits for V), so that we get more bits for V along the length of the stroke, and then we omit the per-particle animation offset (since there are no particles).

PiperOrigin-RevId: 723162303