Remove URI type and BrushProvider interface
Change BrushFamily URI to a client-provided string ID, which has particular meaning only for that client, and may serve as a human-readable ID for debugging etc. If we end up providing a stock brushes interface in the future, will probably do something more in line with what the Jetpack release does with StockBrushes and provide a module interface instead of a URI scheme.

PiperOrigin-RevId: 723928391