Remove URI type and BrushProvider interface
Change BrushFamily URI to a client-provided string ID, which has particular meaning only for that client, and may serve as a human-readable ID for debugging etc. If we end up providing a stock brushes interface in the future, will probably do something more in line with what the Jetpack release does with StockBrushes and provide a module interface instead of a URI scheme.

The ID is referred to as `ClientBrushFamilyId` and similar, to make it clear that the meaning of the ID is client-defined. The `color_texture_id` field of `TextureLayer` is similarly renamed to `client_color_texture_id`.

PiperOrigin-RevId: 724449380