Remove some unused and obsolete CodedMesh proto fields
These were going to be used for serializing some mesh vertex attributes, but weren't general enough to do what we want to do, and that bit was never implemented. Work to implement a more general solution is in progress.

Rename a bunch of fields related to that slightly in response to proto style guidelines / linter.

PiperOrigin-RevId: 725721711