Fix some incorrect declarations in JNI
The parameters passed here are actually the instance object from method calls on Kotlin singletons (and in one case, just a normal instance method on a thing that's _not_ a singleton) not classes from static method calls. jobject and jclass are identical at the C interface, so this will work until you try to pass the incorrectly-labeled jobject to a method that actually that expects a jclass.

PiperOrigin-RevId: 729116090