Handle input modeler correctly when coats are in legacy format
In the legacy format where BrushFamily specifies a single paint and tip instead of a list of coats, we were failing to handle input modeler deserialization. Fix that by ensuring that the same logic is used for all the other fields no matter what format the coats are specified in.

PiperOrigin-RevId: 729252068