Pass array of underlying pointers when constructing BrushPaint and BrushTip via JNI
This is consistent with how BrushFamily and BrushCoat are constructed. It also does a better job with validation. Previously, BrushPaint validated the entire list of TextureLayers each time that was appended, and BrushTip validated the list of BrushBehaviors before that was populated at all. (Though possibly that interface could make harder assumptions that the passed-in instances are already validated / that validation looks at each instance individually.)

In all of those cases, we can be slightly more efficient by releasing the array elements with JNI_ABORT. That frees the buffer of elements, but doesn't try to copy that back to the JVM, which is unnecessary when we're not modifying the elements.

PiperOrigin-RevId: 729493502