Pull JNI logic for vertex/index buffers into InProgressStrokeWrapper
Adjust naming to make it clear when the native code returns a mutable ByteBuffer that should not be mutated (and therefore should be immediately wrapped by asReadOnlyBuffer).

Adjust the functions for casting the native pointer to return only to a single type (InProgressStrokeWrapper&), even though that requires an extra accessor to get the underlying InProgressStroke. Make the usual version for that const, and add a version where the reference is explicitly noted to be mutable, since this adds extra clarity/enforcement about which operations mutate.

PiperOrigin-RevId: 781663877