Fix bug where tiny enough particles caused InProgressStroke handoff to fail
This fixes a bug in `PartitionedMesh::FromMutableMeshGroups` where it could end up with empty outlines, subsequently failing validation in `PartitionedMesh::Data::FromMeshGroups`. The logic in `FromMutableMeshGroups` goes through and converts the overall vertex indices in each outline to pairs of `(mesh index, vertex index)`. In the process, that prunes any stroke-outline vertices that didn't end up in any triangles in the actual mesh (i.e. vertices that were pruned during triangulation). However, if the mesh is trivial enough that it ends up with _no_ triangles, this led to an empty outline. Instead, discard outlines with fewer than three vertices.

For the sake of robustness, also have `BrushTipExtruder::ExtrudeBreakPoint` drop the geometry between two breakpoints if it's too trivial to actually draw (fewer than three vertices). That logic already handled the zero-vertices case, but it's possible to end up with extensions with only two.

While there, fix some logic in `BrushTipExtruder::Extrude` that is supposed to bail out when there are fewer than two non-break extrusions since the previous extrusion. That logic was previously checking two back, but not one back.

PiperOrigin-RevId: 783863387