Partition InProgressStroke in JNI when exceeding 16-bit indices
This does a few things to make the code converting 32-bit to 16-bit indices a bit clearer. First, it casts the data pointer to the correct type and wraps that in a new span instead of multiplying the offset and using `memcpy` to copy a larger region than the pointer size. Second, it uses `resize` instead of `erase` when rewinding to the last complete triangle (generally clearer when erasing a suffix).

PiperOrigin-RevId: 784294384