shearFactor -> skew in Quad
This is consistent with the change for AffineTransform and conventions in Skia and elsewhere. Adjust the documentation to be clearer about how skew works, in particular start with the explanation previously described as "more intuitive". Dropping the "factor", those other conventions (in Android, Skia, etc.) already use "skew" as a ratio (displacement along axis as a multiple of displacement from axis).

PiperOrigin-RevId: 795477700