Enable angle-based upsampling in spring model
This mostly helps with the stroke-normal based projections in the loop contraction mitigation feature. The way that feature works is that the modeled stroke is pulled back towards the corresponding part of the raw input polyline on fast-moving strokes. But this relies on the stroke normal sweeping across the input polyline. When the modeled stroke turns too rapidly in a single step, it can effectively cut off the corner in tight loops.

The way this feature works is that it estimates the absolute change in the angle of the velocity over the input (i.e. the change in the angle of the velocity without upsampling, this is the absolute change and doesn't consider direction of turn so it's limited to a half-turn). Then it upsamples to at least the number obtained by dividing that estimated angle by max_estimated_angle_to_traverse_per_input.

This change in configuration upsamples to at least four points per estimated quarter-turn. This should particularly result in better performance of the loop contraction mitigation on tight corners.

Also, don't needlessly use a forwarding reference for the argument of the lambda passed to std::visit.

PiperOrigin-RevId: 800654472