Simplify logic for managing Skia uniform data
Pull the copy-on-write logic into a single function. Use templated helper functions to handle the sized copy with `memcpy`. Don't handle empty uniform data as a separate case (it's not a common case in Ink, and `SkData::MakeUninitialized` already has special-case logic which returns a singleton empty data in that case).

PiperOrigin-RevId: 804540038