Upsample inputs in `SlidingWindowInputModeler`
Previously, this modeling implementation generated exactly one modeled input per raw input.  Now, it takes a `Duration32 upsampling_period` parameter, and tries to keep the gap between modeled inputs no larger than that, by subdividing the gaps between raw inputs if necessary.

This required a bunch of refactoring to the rest of the modeler to remove the former one-to-one assumption.  In particular, this class was previously a bit loose about referring to the raw inputs as "stable" or "unstable", when really that's a property of the modeled inputs, not the raw inputs.  (In fact, there's no reason why the modeled inputs have to line up elapsed-time-wise with the raw inputs at all, but for now this modeler still always generates a modeled input with the same timestamp as any given raw input.)

PiperOrigin-RevId: 811022568