Add `StrokeInputBatch::Last()` and `First()` methods
Getting the first or last input in a `StrokeInputBatch` turns out to be a pretty common operation.  Getting the last input was previously awkward, requiring writing `inputs.Get(inputs.Size() - 1)`, which we did in a lot of places; `inputs.Last()` is clearer and less error-prone.  Meanwhile, `inputs.First()` isn't really any more convenient than `inputs.Get(0)`, but it reads better, especially when it's next to a corresponding call to `Last()` (which it often is).

PiperOrigin-RevId: 816342480