Bump version for llvm, remove workaround setting up libtinfo5
Bumps a few other module versions to avoid build warnings.

My previous attempt to remove that bit of setup didn't work, and my verification that the workaround could just be removed was thwarted by build caching papering over the problem. Bumping the toolchain dependency should fix it, though, and as a bonus should guarantee a full rebuild in presubmit. Despite the comment there, the issue seems to be with Clang, not gmock (or at least not just). More context in:
https://github.com/bazel-contrib/toolchains_llvm/issues/528

PiperOrigin-RevId: 825598123