Use default input model when decoding a proto with a missing/unknown input model, and remove `ExperimentalRawPositionModel`
`ExperimentalRawPositionModel` was intended to effectively just pass the raw inputs through mostly unchanged, but `ExperimentalNaiveModel` does that in a simpler and more faithful way.  We don't need both of them (and nothing in production should be using either one), so `ExperimentalRawPositionModel` can safely be removed.

Deprecating and removing experimental input models does however create the problem of potentially making older brush files unloadable.  So this change also changes brush proto decoding so that if the input model in the proto is an unrecognized type, the default input modeler is used as a fallback.

PiperOrigin-RevId: 826038799