Use `std::lerp` instead of hand-coded linear interpolation
`std::lerp` has better edge-case properties (e.g. avoiding floating point overflow) than our more naive definitions. We only didn't use it before because it requires C++20, which only recently became available in all of our build environments.

We overload our (internal-only) `Lerp()` function to apply to a bunch of our geometry types, as well as to `float`.  I think its reasonable to still keep and use the `float` version for consistency, but that function now just inline-delegates to `std::lerp`.  Other Ink `Lerp` functions have been updated appropriately.

`SpiralLerp` has been removed entirely, since it was unused.  (It was added in cl/552501553, apparently with the expectation that it would be useful for anti-aliasing, but as far as I can tell it never ended up getting used for that.)

PiperOrigin-RevId: 826613841