Add fuzz tests for stroke input modeling
`StrokeInputModeler` should be able to model any inputs, with any valid input model, and not crash.

Right now, the spring-based modeler (which we are moving away from) rejects some large-value inputs to avoid float overflow, and previously, we were CHECK-failing if that happened. This change removes that CHECK, since we shouldn't be crashing on inputs that pass `StrokeInputBatch` validation.

PiperOrigin-RevId: 827562618