Use modeled input velocity for computing direction-related source values
Previously the brush tip modeler had a rough way to compute velocity, but it makes more sense (and is simpler, and more consistent) to base travel direction on the input modeler's modeled velocity.  We didn't do this before because of shortcomings with the way the spring-based modeler computed velocity, but we've now transitioned over to the sliding window modeler, which doesn't have those issues.

PiperOrigin-RevId: 842789441